Updating to GAMOS 6.0

+++ GamosAnalysis/src/GmEventClassifierByGammaInteraction.cc
@@ -1,28 +1,3 @@
-//
-// ********************************************************************
-// * License and Disclaimer                                           *
-// *                                                                  *
-// * The  GAMOS software  is  copyright of the Copyright  Holders  of *
-// * the GAMOS Collaboration.  It is provided  under  the  terms  and *
-// * conditions of the GAMOS Software License,  included in the  file *
-// * LICENSE and available at  http://fismed.ciemat.es/GAMOS/license .*
-// * These include a list of copyright holders.                       *
-// *                                                                  *
-// * Neither the authors of this software system, nor their employing *
-// * institutes,nor the agencies providing financial support for this *
-// * work  make  any representation or  warranty, express or implied, *
-// * regarding  this  software system or assume any liability for its *
-// * use.  Please see the license in the file  LICENSE  and URL above *
-// * for the full disclaimer and the limitation of liability.         *
-// *                                                                  *
-// * This  code  implementation is the result of  the  scientific and *
-// * technical work of the GAMOS collaboration.                       *
-// * By using,  copying,  modifying or  distributing the software (or *
-// * any work based  on the software)  you  agree  to acknowledge its *
-// * use  in  resulting  scientific  publications,  and indicate your *
-// * acceptance of all terms of the GAMOS Software license.           *
-// ********************************************************************
-//
 
 #include "GmEventClassifierByGammaInteraction.hh"
 #include "GmTrajPoint.hh"
@@ -44,10 +19,10 @@ GmEventClassifierByGammaInteraction::~GmEventClassifierByGammaInteraction()
 }
 
 //---------------------------------------------------------------
-int GmEventClassifierByGammaInteraction::Classify( const std::vector<GmTrajPoint*>& interactionList )
+int64_t GmEventClassifierByGammaInteraction::Classify( const std::vector<GmTrajPoint*>& interactionList )
 {
   //  G4cout << " GmEventClassifierByGammaInteraction::Classify " << interactionList.size() << G4endl;
-  int type = 0;
+  int64_t type = 0;
 
   std::vector<GmTrajPoint*>::const_iterator ite;
   for( ite = interactionList.begin(); ite != interactionList.end(); ite++ ){ 
@@ -73,14 +48,14 @@ void GmEventClassifierByGammaInteraction::BuildProcessNames()
 }
 
 //---------------------------------------------------------------
-int GmEventClassifierByGammaInteraction::Classify( const std::vector<GmTrajStep*>& interactionList )
+int64_t GmEventClassifierByGammaInteraction::Classify( const std::vector<GmTrajStep*>& interactionList )
 {
   //  G4cout << " GmEventClassifierByGammaInteraction::Classify " << interactionList.size() << G4endl;
   if( !bProcessNamesBuilt ) {
     BuildProcessNames();
   }
 
-  int type = 0;
+  int64_t type = 0;
 
   std::vector<GmTrajStep*>::const_iterator ite;
   for( ite = interactionList.begin(); ite != interactionList.end(); ite++ ){ 
@@ -92,9 +67,9 @@ int GmEventClassifierByGammaInteraction::Classify( const std::vector<GmTrajStep*
 }
 
 //---------------------------------------------------------------
-int GmEventClassifierByGammaInteraction::Classify1Interaction( const G4String& processName )
+int64_t GmEventClassifierByGammaInteraction::Classify1Interaction( const G4String& processName )
 {
-  int type = 0;
+  int64_t type = 0;
 
   if( processName == thePhotoElecName ){ 
     type = 10000;
@@ -127,9 +102,8 @@ int GmEventClassifierByGammaInteraction::Classify1Interaction( const G4String& p
 }
 
 //---------------------------------------------------------------
-G4String GmEventClassifierByGammaInteraction::GetClassificationName( int ic ) 
+G4String GmEventClassifierByGammaInteraction::GetClassificationName( int64_t ic ) 
 {
-
   G4String name = "Evt:";
 
   int iPE = ic/10000;
